requirements.txt
django-cloudinary-storage[video]
django-utils-six==2.0
cloudinary==1.17.0
django-cloudinary-storage==0.2.3
django-background-tasks==1.2.5
aioredis==1.3.0
appdirs==1.4.3
django_crontab==0.7.1
APScheduler==3.6.3
asgiref==3.2.2
asn1crypto==0.24.0
astroid==2.4.1
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
Babel==2.8.0
beautifulsoup4==4.9.1
certifi==2019.6.16
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
colorama==0.4.3
configparser==3.7.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.41.0
django-apscheduler==0.3.0
django-cors-headers==3.1.0
django-extensions==2.2.9
django-heroku==0.3.1
django-mobile-app-distribution==0.5
django-model-utils==4.0.0
django-multiselectfield==0.1.12
django-mysql==3.5.0
django-notifications-hq==1.6.0
django-oauth-toolkit==1.3.2
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-rest-multiple-models==2.1.3
django-rest-passwordreset==1.1.0
django-restframework==0.0.1
django-social-auth==0.7.28
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==4.3.0
docutils==0.16
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
hiredis==1.0.0
httplib2==0.17.3
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.6.0
incremental==17.5.0
isort==4.3.21
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
MouseInfo==0.1.3
msgpack==0.6.2
numpy==1.17.0
oauth2==1.9.0.post1
oauthlib==3.0.2
packaging==20.4
pandas==0.25.0
phonenumbers==8.12.2
Pillow==6.2.0
pipenv==2018.11.26
pluggy==0.13.1
psycopg2==2.8.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pycparser==2.19
PyGetWindow==0.0.8
Pygments==2.6.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.5.2
PyMsgBox==1.0.8
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.0
python-decouple==3.3    
python-openid==2.2.5
python3-openid==3.1.0
PyTweening==1.0.3
pytz==2019.2
pywhatkit==2.1
requests==2.22.0
requests-oauthlib==1.2.0
service-identity==18.1.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==2.0.1
spepy==1.0
sqlparse==0.3.0
swapper==1.1.2.post1
toml==0.10.0
tox==3.15.1
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.3
uuid==1.30
virtualenv==20.0.20
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
whitenoise==5.1.0
wikipedia==1.4.0
wrapt==1.12.1
zipp==3.1.0
zope.interface==4.6.0
django-heroku==0.3.1